Creation of Deck Data and Implementation of Parser
We have Deck Data class that represent raw data of a deck.
We also now have a custom parser that reads in the deck data from files.
And we also can now out put to files and even output streams to output that deck.
It can also save that deck provided a filename to use.