Deck Component All Set Up
All set up,
Waits for card events like draw add discard, peek, and a shuffle event.
Produces card drawn, discarded, or peeked with event data that holds which card and what placement it was on relative to top of pool.
and also has a chance of queueing the Deck Empty event which means there are no more cards in the deck, which happens purely before any attempt to get another card from the deck happens.