Initial setup of just the boxes.
This is just the start,
It puts down two buttons
one for a new game, and one for a deck builder.
Observed is that texboxes appear underneath it.
The textbox has to separated sadly. no way to properly set it up. but tis ok.
UIClickables will be added later when they are complete.