Added fix for networked Blast ability
Issue was when Blast was activated, it would send the AbilityPacket, then the target would get destroyed, then the host ServerGame received the packet and tried to access the destroyed target.
Fixed by converting the AbilityPacket vectors to hold only primitives and only get the actual objects when getters are called.