Context Menu Behaviour begun.
Row Control:

Added data structure for organizing elements that exist within a context menu. This will be used for ordering the elements in the menu to rendered in the desired positions.

Next step. Add a method for ordering elements using their position and 2d data to be rendered as desired within the menu.