Button Added
The top right corner of the screen now has the turn end button on it for player convieniece. Might be worth removing the end turn button from the context list of abilities for simplicity.

Also: The turn counter is covered by the button so I will need to find a place to put it. Ill take care of this at some point.