Finished Ambient Particles
Particles will now appear periodically on key-tiles as well as in the surrounding area around the map.

1- Desert Sand will produce dust that travels in the wind

2- Swamp will produce gas

3- Water will twinkle

4- Flies will appear and fly from right to left.

Next step: I want to write some new effectors so that it will rain once in a while for an extended period of time and do some other cool stuff like better flight animations for the flies, as well as stuff that circles around a position.